+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning][Semantic Versioning].
 - Added `HasError` and `ErrorMessage` pattern to enable consistent error reporting from handlers to views.
 - Added an article about DTO models for future documentation 
 - AddAlertPopup now pops up to let user add name or county to alerts in configview
+- `AlertHelper` to find the County or patient number in `global_config.json` and show user an alert
 
 ### Changed